Factor out errors into own file and write some tests
